app/requirements.txt
Django==2.0.2
redis==2.10.3
django-redis==4.3.0
django-registration==2.4.1
djangorestframework==3.7.7
gunicorn==19.4.5
mysqlclient==1.3.12
configparser==3.5.0
python-dateutil
djangorestframework==3.7.7
social-auth-app-django==3.1.0
requests==2.18.4
PyYAML==3.13
ua-parser==0.8.0
user-agents==1.1.0
django-user-agents==0.3.2
django-mongolog==0.9.4
pymongo==3.7.2
rest-social-auth==2.1.0
